Update bubble_sort_cmdLn.c
There are 3 mains that serves different purposes. The first main is for the actual sort function; the second main is just a simple main to return all given args to stdout; and the third main is a more elaborate form of the second main.
To run this program, kindly comment out the mains you are not intending to use then compile and run this via your own editor.