Update bubble_sort_cmdLn.c
Bubble_sort algorithm sorts by repeatedly stepping through the list, compares adjacent elements, and swaps them if they are in the wrong order.
There are 3 mains that serves different purposes. The first main is for the actual sort function; the second main is just a simple main to return all given args to stdout; and the third main is a more elaborate form of the second main.
To run this program, kindly comment out the mains you are not intending to use then compile and run this via your own editor.